+++ example/android/app/src/main/kotlin/com/example/internet_check_in_app_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.internet_check_in_app_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()