+++ lib/bloc/connectivity_cubit.dart
@@ -1,0 +1,17 @@
+import 'package:internet_check_in_app/internet_check_in_app.dart';
+
+enum ConnectivityStatus { connected, disconnected }
+
+class ConnectivityCubit extends Cubit<ConnectivityStatus> {
+  final Connectivity _connectivity = Connectivity();
+
+  ConnectivityCubit() : super(ConnectivityStatus.connected) {
+    _connectivity.onConnectivityChanged.listen((result) {
+      if (result == ConnectivityResult.none) {
+        emit(ConnectivityStatus.disconnected);
+      } else {
+        emit(ConnectivityStatus.connected);
+      }
+    });
+  }
+}