+++ lib/ui/no_internet_page.dart
@@ -1,0 +1,30 @@
+import 'package:internet_check_in_app/internet_check_in_app.dart';
+
+class NoInternetScreen extends StatelessWidget {
+  const NoInternetScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return WillPopScope(
+      onWillPop: () async => false,
+      child: Scaffold(
+        backgroundColor: Colors.white,
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Image.asset('assets/connection.png',package: 'internet_check_in_app',),
+              SizedBox(height: 20),
+              Text(
+                "No Internet Connection",
+                style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+              ),
+              SizedBox(height: 10),
+              Text("Please check your network settings."),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}